Define new gRPC module to handle gRPC services import
