--- apps/api/Dockerfile.dev
@@ -1,16 +1,0 @@
-FROM node:lts-alpine AS base
-ENV PNPM_HOME="/pnpm"
-ENV PATH="$PNPM_HOME:$PATH"
-RUN corepack enable
-WORKDIR /app
-
-FROM base AS deps
-COPY package*.json pnpm*.yaml ./
-RUN --mount=type=cache,id=pnpm,target=/pnpm/store pnpm install
-
-FROM base
-COPY . .
-COPY --from=deps /app/node_modules ./node_modules
-
-EXPOSE 3000
-CMD ["pnpm", "start:debug", "api"]
