--- apps/api/src/main.ts
@@ -1,8 +1,0 @@
-import { NestFactory } from '@nestjs/core';
-import { ApiModule } from './api.module';
-
-async function bootstrap() {
-  const app = await NestFactory.create(ApiModule);
-  await app.listen(process.env.port ?? 3000);
-}
-bootstrap();
