--- apps/api/src/payments/dtos/create-checkout-session.dto.ts
@@ -1,3 +1,0 @@
-export class CreateCheckoutSessionDto {
-  plans: string[];
-}
