+++ apps/main/src/main.module.ts
@@ -1,10 +1,8 @@
+import { AuthenticationModule, GrpcModule } from '@common';
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { ClientsModule, Transport } from '@nestjs/microservices';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { SERVICES_PACKAGE_NAME } from '@protos/resources.service';
 import * as Joi from 'joi';
-import { resolve } from 'node:path';
 import { ResourcesController } from './resources/resources.controller';
 
 @Module({
@@ -27,23 +25,8 @@ import { ResourcesController } from './resources/resources.controller';
         synchronize: true,
       }),
     }),
-
-    ClientsModule.registerAsync([
-      {
-        name: SERVICES_PACKAGE_NAME,
-        inject: [ConfigService],
-        useFactory: (configService: ConfigService) => ({
-          transport: Transport.GRPC,
-          options: {
-            url: configService.getOrThrow<string>('SERVICES_URL'),
-            package: SERVICES_PACKAGE_NAME,
-            protoPath: [
-              resolve(__dirname, '../../../protos/src/resources.service.proto'),
-            ],
-          },
-        }),
-      },
-    ]),
+    GrpcModule,
+    AuthenticationModule,
   ],
   controllers: [ResourcesController],
   providers: [],
