+++ apps/main/src/resources/resources.controller.ts
@@ -1,18 +1,26 @@
-import { Body, Controller, Inject, OnModuleInit, Post } from '@nestjs/common';
+import { GRPC_SERVICES, JwtAuthGuard } from '@common';
+import {
+  Body,
+  Controller,
+  Inject,
+  OnModuleInit,
+  Post,
+  UseGuards,
+} from '@nestjs/common';
 import { ClientGrpc } from '@nestjs/microservices';
 import {
   CreateResourceRequest,
   RESOURCES_SERVICE_NAME,
   ResourcesServiceClient,
-  SERVICES_PACKAGE_NAME,
 } from '@protos/resources.service';
 
 @Controller('resources')
+@UseGuards(JwtAuthGuard)
 export class ResourcesController implements OnModuleInit {
   private resourcesServiceClient: ResourcesServiceClient;
 
   constructor(
-    @Inject(SERVICES_PACKAGE_NAME)
+    @Inject(GRPC_SERVICES)
     private readonly client: ClientGrpc,
   ) {}
