+++ apps/services/src/authentication/authentication.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { AuthenticationController } from './authentication.controller';
+
+describe('AuthenticationController', () => {
+  let controller: AuthenticationController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [AuthenticationController],
+    }).compile();
+
+    controller = module.get<AuthenticationController>(AuthenticationController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
