+++ apps/services/src/authentication/authentication.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { AuthenticationService } from './authentication.service';
+import { AuthenticationController } from './authentication.controller';
 
 @Module({
   imports: [],
-  controllers: [],
+  controllers: [AuthenticationController],
   providers: [AuthenticationService],
   exports: [AuthenticationService],
 })
