+++ apps/services/src/common/guards/jwt-auth.guard.ts
@@ -2,43 +2,22 @@ import { JwtPayload } from '@clerk/types';
 import {
   CanActivate,
   ExecutionContext,
-  Inject,
   Injectable,
   Logger,
-  OnModuleInit,
   UnauthorizedException,
 } from '@nestjs/common';
-import { ClientGrpc } from '@nestjs/microservices';
-import {
-  AUTHENTICATION_SERVICE_NAME,
-  AuthenticationServiceClient,
-  SERVICES_PACKAGE_NAME,
-} from '@protos/authentication.service';
-import { firstValueFrom } from 'rxjs';
+import { AuthenticationService } from '../../authentication/authentication.service';
 
 @Injectable()
-export class JwtAuthGuard implements CanActivate, OnModuleInit {
+export class JwtAuthGuard implements CanActivate {
   private readonly logger = new Logger(JwtAuthGuard.name);
-  private authenticationServiceClient: AuthenticationServiceClient;
 
-  constructor(
-    @Inject(SERVICES_PACKAGE_NAME)
-    private readonly client: ClientGrpc,
-  ) {}
-
-  onModuleInit() {
-    this.authenticationServiceClient =
-      this.client.getService<AuthenticationServiceClient>(
-        AUTHENTICATION_SERVICE_NAME,
-      );
-  }
+  constructor(private readonly authenticationService: AuthenticationService) {}
 
   async canActivate(ctx: ExecutionContext): Promise<boolean> {
     try {
-      const jwt = await firstValueFrom(
-        this.authenticationServiceClient.verify({
-          token: this.getAuthorization(ctx),
-        }),
+      const jwt = await this.authenticationService.verify(
+        this.getAuthorization(ctx),
       );
 
       this.addJWT(ctx, jwt);
@@ -59,22 +38,12 @@ export class JwtAuthGuard implements CanActivate, OnModuleInit {
       );
     }
 
-    if (ctx.getType() === 'rpc') {
-      const req = ctx.switchToRpc().getData();
-      return req.jwt;
-    }
-
     throw new Error(`Unsupported ${ctx.getType()} request type`);
   }
 
   private addJWT(ctx: ExecutionContext, jwt: JwtPayload): void {
     if (ctx.getType() === 'http') {
       ctx.switchToHttp().getRequest().jwt = jwt;
-      return;
-    }
-
-    if (ctx.getType() === 'rpc') {
-      ctx.switchToRpc().getData().jwt = jwt;
       return;
     }
