+++ apps/services/src/payments/payments.service.ts
@@ -1,6 +1,5 @@
-import { Inject, Injectable, Logger, RawBodyRequest } from '@nestjs/common';
+import { Inject, Injectable, Logger } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { Request } from 'express';
 import Stripe from 'stripe';
 import { AuthenticationService } from '../authentication/authentication.service';
 import { STRIPE_CLIENT } from '../clients';
@@ -106,34 +105,5 @@ export class PaymentsService {
     });
 
     return session;
-  }
-
-  async processEvent(request: RawBodyRequest<Request>): Promise<void> {
-    let event = request.body as Stripe.Event;
-
-    const endpointSecret = this.configService.getOrThrow<string>(
-      'STRIPE_ENDPOINT_SECRET',
-    );
-
-    if (!endpointSecret) {
-      try {
-        const signature = request.header('stripe-signature');
-        event = this.stripeClient.webhooks.constructEvent(
-          request.body,
-          signature,
-          endpointSecret,
-        );
-      } catch (err) {
-        this.logger.error(
-          '⚠️ Webhook signature verification failed.',
-          err.message,
-        );
-        throw new Error('⚠️ Webhook signature verification failed.');
-      }
-    }
-
-    await Promise.all(
-      this.eventHandlers.map((handler) => handler.handle(event)),
-    );
   }
 }
