+++ apps/services/src/users/users.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
 import { AuthenticationModule } from '../authentication/authentication.module';
-import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
+import { WebhooksController } from './webhooks/webhooks.controller';
+import { WebhooksService } from './webhooks/webhooks.service';
 
 @Module({
   imports: [AuthenticationModule],
-  providers: [UsersService],
-  controllers: [UsersController],
+  providers: [UsersService, WebhooksService],
+  controllers: [WebhooksController],
   exports: [UsersService],
 })
 export class UsersModule {}
