+++ docker-compose.yml
@@ -1,28 +1,6 @@
 name: saas-template
 
 services:
-  st-api:
-    container_name: st-api
-    depends_on:
-      - st-main
-      - st-services
-    build:
-      context: .
-      dockerfile: ./apps/api/Dockerfile.dev
-    command: pnpm start:debug api
-    env_file:
-      - ./apps/api/.env
-    environment:
-      RABBIT_MQ_URL: "amqp://admin:password@st-rabbitmq/"
-    volumes:
-      - ./apps/api:/app/apps/api
-      - ./libs:/app/libs
-      - ./protos:/app/protos
-    ports:
-      - 9231:9229
-    networks:
-      - st-network
-
   st-services:
     container_name: st-services
     depends_on:
