+++ libs/common/src/authentication/authentication.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { GrpcModule } from '../grpc/grpc.module';
+
+@Module({
+  imports: [GrpcModule],
+})
+export class AuthenticationModule {}
