+++ libs/common/src/grpc/grpc.services.ts
@@ -1,0 +1,1 @@
+export const GRPC_SERVICES = Symbol('GRPC_SERVICES');
