+++ libs/common/src/grpc/index.ts
@@ -1,0 +1,3 @@
+export * from './grpc.module';
+export * from './grpc.services';
+export * from './grpc.utils';
