+++ libs/common/src/index.ts
@@ -1,0 +1,3 @@
+export * from './authentication/authentication.module';
+export * from './authentication/guards/jwt-auth.guard';
+export * from './grpc';
