+++ libs/protos/src/utils/proto-message.utils.ts
@@ -1,0 +1,8 @@
+export class ProtoMessageUtils {
+  static toMap(obj: { [key: string]: unknown }): { [key: string]: string } {
+    return Object.entries(obj).reduce(
+      (acc, [key, value]) => ({ ...acc, [key]: JSON.stringify(value) }),
+      {},
+    );
+  }
+}
