+++ nginx.conf
@@ -3,8 +3,12 @@ http {
   # The upstream conatiner specficiation,
   # based on docs, to be reacheded is:
   # container_name:port
-  upstream api {
-    server st-api:3000;
+  upstream main {
+    server st-main:3000;
+  }
+
+  upstream services {
+    server st-services:3000;
   }
 
   server {
@@ -18,7 +22,7 @@ http {
     proxy_send_timeout 30s;
 
     location / {
-      proxy_pass http://api/;
+      proxy_pass http://main/;
     }
   }
 }
