+++ protos/src/services/resources.service.proto
